demos/requirements.txt
strands-agents
strands-agents[a2a]
strands-agents-tools[a2a_client]
strands-agents-tools
